+++ subsys/storage/flash_map/flash_map.c
@@ -10,12 +10,12 @@
 #include <zephyr/types.h>
 #include <stddef.h>
 #include <sys/types.h>
-#include <device.h>
-#include <storage/flash_map.h>
+#include <zephyr/device.h>
+#include <zephyr/storage/flash_map.h>
 #include "flash_map_priv.h"
-#include <drivers/flash.h>
+#include <zephyr/drivers/flash.h>
 #include <soc.h>
-#include <init.h>
+#include <zephyr/init.h>
 
 void flash_area_foreach(flash_area_cb_t user_cb, void *user_data)
 {
@@ -93,7 +93,7 @@ int flash_area_erase(const struct flash_area *fa, off_t off, size_t len)
 	return rc;
 }
 
-uint8_t flash_area_align(const struct flash_area *fa)
+uint32_t flash_area_align(const struct flash_area *fa)
 {
 	const struct device *dev;
